start removing the "new" MJ API, commit RedBuildCraft engine implementation

+++ api/buildcraft/api/power/ILaserTarget.java
@@ -33,9 +33,9 @@ public interface ILaserTarget {
 	 */
 	boolean isInvalidTarget();
 
-	double getXCoord();
+	int getXCoord();
 
-	double getYCoord();
+	int getYCoord();
 
-	double getZCoord();
+	int getZCoord();
 }
